feat: 로또를 번호를 생성하는 LottoGenerator 클래스 구현
